+++ src/test/java/lotto/domain/LottoIssuanceCountTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.exception.ErrorMessage;
+import lotto.exception.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
