+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.exception.ErrorMessage;
+import lotto.exception.ErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
